Use vim-plugin in README, fzf Ag update
Previously, when using *Ag* with fzf it was not possible to select
multiple files. Now, multiple files can be selected and comes in a
quickfix window. Not a perfect selection, but still its okay.

With the *vim-hopping* plugin, the search was cool, but didn't know on
how to move between results, this commit also has the setting for that.
Thanks to osymanga for replying github.

Also updated the README that was bit out of date with NeoBundle. Have
updated with *vim-plug* installation. So far liking vim-plug pretty good